requirements_2.7.txt
configobj==5.0.6
psutil==5.2.2
python-statsd==2.1.0
